REFACTOR backend : add missing types

+++ apps/backend/src/timers/execute-action.timer.ts
@@ -9,11 +9,15 @@ export default class ExecuteActionTimer extends BaseTimer {
 	DECREMENT_CREDIT_MESSAGE:string = 'decrement-credit';
 	REMOVE_ACTION_MESSAGE:string = 'remove-action';
 
+	private creditService:CreditService;
+	private actionService:ActionService;
 	constructor(
-		@inject(CreditService)private creditService:CreditService,
-        @inject(ActionService)private actionService:ActionService,	
+		@inject(CreditService)creditService:CreditService,
+        @inject(ActionService)actionService:ActionService,	
 	){
 		super();
+		this.creditService = creditService;
+		this.actionService = actionService;
 	}
 
 	execute = async ():Promise<void> => {
@@ -31,13 +35,16 @@ export default class ExecuteActionTimer extends BaseTimer {
 			{ $limit: 1 } 
 		  ]);
 
-        if (action.length > 0){
-            const removedAction = await this.actionService.delete(action[0]._id) 
-			io.emit(this.REMOVE_ACTION_MESSAGE.toString(), JSON.stringify(removedAction) );
-
-			const credit = await this.creditService.getById(removedAction.credit)
-			const updatedcredit = await this.creditService.decrementCredit(credit)	
-			io.emit(this.DECREMENT_CREDIT_MESSAGE.toString(), JSON.stringify(updatedcredit) );
+		if (action && action.length > 0){
+			const removedAction = await this.actionService.delete(action[0]._id) 
+			if (removedAction) {
+				io.emit(this.REMOVE_ACTION_MESSAGE.toString(), JSON.stringify(removedAction) );
+				const credit = await this.creditService.getById(removedAction.credit.toString())
+				if (credit) {
+					const updatedcredit = await this.creditService.decrementCredit(credit)	
+					io.emit(this.DECREMENT_CREDIT_MESSAGE.toString(), JSON.stringify(updatedcredit) );
+				}
+			}
 		}
         
 	}
